Validator: moved units to measurement package, fixed Validator tests with new thresholds, modified build_validator to receive Quantities.
